Refactor: split lapor, tambah, & setalias controllers
